into oXc_kL
Conflicts:
	src/Engine/State.cpp
	src/Engine/Surface.cpp
	src/Engine/Surface.h
	src/Geoscape/DogfightState.cpp
	src/Geoscape/GeoscapeState.cpp
	src/Interface/BattlescapeButton.cpp
	src/Interface/BattlescapeButton.h
	src/Interface/TextButton.cpp
	src/Interface/TextButton.h
	src/Ruleset/RuleGlobe.cpp

+++ src/Ruleset/RuleGlobe.cpp
@@ -162,13 +162,13 @@ void RuleGlobe::load(const YAML::Node& node)
 		}
 	}
 
-	Globe::COUNTRY_LABEL_COLOR	= node["countryColor"]	.as<Uint8>(Globe::COUNTRY_LABEL_COLOR);
-	Globe::CITY_LABEL_COLOR		= node["cityColor"]		.as<Uint8>(Globe::CITY_LABEL_COLOR);
-	Globe::BASE_LABEL_COLOR		= node["baseColor"]		.as<Uint8>(Globe::BASE_LABEL_COLOR);
-	Globe::LINE_COLOR			= node["lineColor"]		.as<Uint8>(Globe::LINE_COLOR);
+	Globe::COUNTRY_LABEL_COLOR	= static_cast<Uint8>(node["countryColor"]	.as<int>(Globe::COUNTRY_LABEL_COLOR));
+	Globe::CITY_LABEL_COLOR		= static_cast<Uint8>(node["cityColor"]		.as<int>(Globe::CITY_LABEL_COLOR));
+	Globe::BASE_LABEL_COLOR		= static_cast<Uint8>(node["baseColor"]		.as<int>(Globe::BASE_LABEL_COLOR));
+	Globe::LINE_COLOR			= static_cast<Uint8>(node["lineColor"]		.as<int>(Globe::LINE_COLOR));
 
 	if (node["oceanPalette"])
-		Globe::OCEAN_COLOR = Palette::blockOffset(node["oceanPalette"].as<Uint8>(12));
+		Globe::OCEAN_COLOR = Palette::blockOffset(static_cast<Uint8>(node["oceanPalette"].as<int>(12)));
 }
 
 /**
